requirements.txt
allure-pytest==2.8.32
pytest==6.2.1
pytest-rerunfailures==9.1.1
requests==2.25.0
selenium==3.141.0